Remove windows code and simplify linux
Windows will not use containerd and its just unused code and unneed
complexity to keep it all around.
