--- ctr/container_windows.go
@@ -1,6 +1,0 @@
-package main
-
-// TODO Windows: This will have a very different implementation
-func createStdio() (s stdio, err error) {
-	return stdio{}, nil
-}
